+++ sdk/lib/_internal/vm/bin/process_patch.dart
@@ -534,19 +534,16 @@ class _ProcessImpl extends _ProcessImplNativeWrapper implements Process {
   _wait(_NativeSocket stdin, _NativeSocket stdout, _NativeSocket stderr,
       _NativeSocket exitHandler) native "Process_Wait";
 
-  Stream<List<int>> get stdout {
-    return _stdout;
-  }
+  Stream<List<int>> get stdout =>
+      _stdout ?? (throw StateError("Process is detached"));
 
-  Stream<List<int>> get stderr {
-    return _stderr;
-  }
+  Stream<List<int>> get stderr =>
+      _stderr ?? (throw StateError("Process is detached"));
 
-  IOSink get stdin {
-    return _stdin;
-  }
+  IOSink get stdin => _stdin ?? (throw StateError("Process is detached"));
 
-  Future<int> get exitCode => _exitCode != null ? _exitCode.future : null;
+  Future<int> get exitCode =>
+      _exitCode?.future ?? (throw StateError("Process is detached"));
 
   bool kill([ProcessSignal signal = ProcessSignal.sigterm]) {
     if (signal is! ProcessSignal) {